refactor(custom-hook): separate mouse event from shapeLeaf
